feat(yk-color-picker): refactor click and key events handling
This commit optimizes event-handling code in the 'YKColorPicker' class. We replaced an if-else condition with a ternary operator in the '_onClickTarget' method for cleaner, more readable code. In the '_onKeyUpClose' method, we added early exit conditions to prevent keyup event propagation when target event is option's target when Enter key is pressed as well as destructure 'event' object for better readability.